Delete nodes module
