--- thermostat/nodes/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Device nodes that we can use and deploy locally."""
